feat: convert remaining timeline endpoints
