+++ Backend/src/Modules/Timelines/Timelines.Api/Controllers/Timelines/TimelinesController.cs
@@ -1,17 +1,19 @@
 using System;
 using System.Threading.Tasks;
 using BuildingBlocks.Application.Pagination;
-using BuildingBlocks.Domain.Timelines.Timeline.Dtos;
 using BuildingBlocks.Domain.Timelines.Timeline.ValueObjects;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using OpenIddict.Validation.AspNetCore;
 using Timelines.Application.Entities.Timelines.Commands.CreateTimeline;
 using Timelines.Application.Entities.Timelines.Commands.CreateTimelineWithTemplate;
+using Timelines.Application.Entities.Timelines.Commands.DeleteTimeline;
+using Timelines.Application.Entities.Timelines.Commands.UpdateTimeline;
 using Timelines.Application.Entities.Timelines.Queries.GetTimelineById;
+using Timelines.Application.Entities.Timelines.Queries.ListNodesByTimelineId;
 using Timelines.Application.Entities.Timelines.Queries.ListTimelines;
 
-namespace Timelines.Api.Controllers;
+namespace Timelines.Api.Controllers.Timelines;
 
 [ApiController]
 [Route("api/[controller]")]
@@ -42,6 +44,18 @@ public class TimelinesController(ISender sender) : ControllerBase
             return NotFound();
 
         var response = result.Adapt<GetTimelineByIdResponse>();
+        return Ok(response);
+    }
+
+    [HttpGet("{timelineId}/Nodes")]
+    [ProducesResponseType(typeof(ListNodesByTimelineIdResponse), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    public async Task<ActionResult<ListNodesByTimelineIdResponse>> ListNodesByTimelineId(
+        [FromRoute] string timelineId,
+        [FromQuery] PaginationRequest query)
+    {
+        var result = await sender.Send(new ListNodesByTimelineIdQuery(timelineId, query));
+        var response = result.Adapt<ListNodesByTimelineIdResponse>();
         return Ok(response);
     }
 
@@ -82,24 +96,36 @@ public class TimelinesController(ISender sender) : ControllerBase
             new { timelineId = response.Timeline.Id },
             response);
     }
+
+    [HttpPut("{timelineId}")]
+    [ProducesResponseType(typeof(UpdateTimelineResponse), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    public async Task<ActionResult<UpdateTimelineResponse>> Update(
+        [FromRoute] string timelineId,
+        [FromBody] UpdateTimelineRequest request)
+    {
+        var command = new UpdateTimelineCommand
+        {
+            Id = TimelineId.Of(Guid.Parse(timelineId)),
+            Title = request.Title,
+            Description = request.Description
+        };
+        var result = await sender.Send(command);
+        var response = result.Adapt<UpdateTimelineResponse>();
+
+        return Ok(response);
+    }
+    
+    [HttpDelete("{timelineId}")]
+    [ProducesResponseType(typeof(DeleteTimelineResponse), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    public async Task<ActionResult<DeleteTimelineResponse>> Delete(
+        [FromRoute] string timelineId)
+    {
+        var result = await sender.Send(new DeleteTimelineCommand(timelineId));
+        var response = result.Adapt<DeleteTimelineResponse>();
+
+        return Ok(response);
+    }
 }
-
-public record ListTimelinesResponse(PaginatedResult<TimelineDto> Timelines);
-
-public record GetTimelineByIdResponse(TimelineDto Timeline);
-
-public class CreateTimelineRequest
-{
-    public string Title { get; set; }
-    public string Description { get; set; }
-}
-
-public record CreateTimelineResponse(TimelineId Id);
-
-public class CreateTimelineWithTemplateRequest
-{
-    public string Title { get; set; }
-    public string Description { get; set; }
-}
-
-public record CreateTimelineWithTemplateResponse(TimelineBaseDto Timeline);
