+++ Backend/src/Modules/Timelines/Timelines.Api/Controllers/Timelines/TimelinesRequests.cs
@@ -1,0 +1,19 @@
+namespace Timelines.Api.Controllers.Timelines;
+
+public class CreateTimelineRequest
+{
+    public string Title { get; set; }
+    public string Description { get; set; }
+}
+
+public class CreateTimelineWithTemplateRequest
+{
+    public string Title { get; set; }
+    public string Description { get; set; }
+}
+
+public class UpdateTimelineRequest
+{
+    public string Title { get; set; }
+    public string Description { get; set; }
+}
