+++ Backend/src/Modules/Timelines/Timelines.Api/Extensions/ServiceCollectionExtensions.cs
@@ -2,7 +2,7 @@ using BuildingBlocks.Api.Converters;
 using BuildingBlocks.Application.Data;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Timelines.Api.Controllers;
+using Timelines.Api.Controllers.Timelines;
 using Timelines.Application.Data;
 using Timelines.Application.Extensions;
 using Timelines.Infrastructure.Data.Extensions;
@@ -24,7 +24,7 @@ public static class ServiceCollectionExtensions
     private static IServiceCollection AddApiServices(this IServiceCollection services)
     {
         TypeAdapterConfig.GlobalSettings.Scan(typeof(TimelineIdConverter).Assembly);
-        
+
         services.AddControllers()
             .AddApplicationPart(typeof(TimelinesController).Assembly);
